refactor(importer): importer functions should always call a callback instead of returning a Promise.
BREAKING CHANGE: Compatibility with node-sass: importers so far had to either call a callback (when they're called in the rendering phase, by `node-sass`) or return a Promise (when they're called in the extracting phase, by `makeImporter()`). From now, importers should always call the callback function provided in the third parameter of the importer function.

fix #36